Put componenet into own .tsx and .scss files
I created FeatureCalloutWidget.tsx and FeatureCalloutWidget.scss
