+++ client/src/containers/App.tsx
@@ -7,6 +7,7 @@ import "styles/App.css";
 import ScrollToTop from "../components/ScrollToTop";
 import SocialShare from "../components/SocialShare";
 import Modal from "../components/Modal";
+import FeatureCalloutWidget from "../components/FeatureCalloutWidget";
 
 // import top-level containers (i.e. pages)
 import { I18n, LocaleNavLink, LocaleLink as Link, LocaleSwitcher } from "../i18n";
@@ -14,10 +15,7 @@ import { withI18n, withI18nProps } from "@lingui/react";
 import { createWhoOwnsWhatRoutePaths } from "../routes";
 import { VersionUpgrader } from "./VersionUpgrader";
 import { useMachine } from "@xstate/react";
-import Timeline_gif from "../assets/img/Feature_callout_gifs/Timeline.gif"
-import Spanish_gif from "../assets/img/Feature_callout_gifs/Spanish.gif"
-import URLS_gif from "../assets/img/Feature_callout_gifs/URLS.gif"
-import LastSold_gif from "../assets/img/Feature_callout_gifs/LastSold.gif"
+
 import HomePage from "./HomePage";
 import AddressPage from "./AddressPage";
 import BBLPage from "./BBLPage";
@@ -34,8 +32,6 @@ type Props = {};
 
 type State = {
   showEngageModal: boolean;
-  entryIndex: number;
-  entries: { index: string; title: string; description: string; img: string }[];
 };
 
 const HomeLink = withI18n()((props: withI18nProps) => {
@@ -97,68 +93,9 @@ export default class App extends Component<Props, State> {
 
     this.state = {
       showEngageModal: false,
-      entryIndex: 0,
-      entries: [
-        {
-          index: "1 of 4",
-          title: "Timeline Tab",
-          description: "Click the Timeline tab to view info about your building over time.",
-          img: Timeline_gif,
-        },
-        {
-          index: "2 of 4",
-          title: "Spanish Support",
-          description:
-            "Who Owns What is now available in Spanish. Click “ES” in the upper right corner to switch your language.",
-          img: Spanish_gif,
-        },
-        {
-          index: "3 of 4",
-          title: "Unique Tab URLs",
-          description:
-            "It's now possible to share links to specific tabs (Overview, Timeline, Portfolio, & Summary).",
-          img: URLS_gif,
-        },
-        {
-          index: "4 of 4",
-          title: "Last Sold",
-          description: "The Overview and Portfolio tabs now display the date and price from when your building was last sold.",
-          img: LastSold_gif,
-        },
-      ],
     };
   }
 
-  nextEntry() {
-    this.state.entryIndex === this.state.entries.length - 1
-      ? this.setState({ entryIndex: 0 })
-      : this.setState({ entryIndex: this.state.entryIndex + 1 });
-  }
-
-  prevEntry() {
-    this.state.entryIndex === 0
-      ? this.setState({ entryIndex: this.state.entries.length - 1 })
-      : this.setState({ entryIndex: this.state.entryIndex - 1 });
-  }
-
-  toggleWidget(widget: any) {
-    let tooltipTriangle = document.querySelector(".widget-tooltip-triangle");
-    let infoButton = document.querySelector(".widget-button-info");
-    if (widget.style.display === "none") {
-      widget.style.display = "inline-block";
-      if (tooltipTriangle) tooltipTriangle.classList.add("toggled");
-      if (infoButton) infoButton.classList.add("pressed");
-    } else {
-      widget.style.display = "none";
-      if (tooltipTriangle) tooltipTriangle.classList.remove("toggled");
-      if (infoButton) infoButton.classList.remove("pressed");
-    }
-  }
-
-  
-
-  
- // ReactDOM.render(<ToggleInfoButton />, document.querySelector("#info-button-container"))
   
 
   render() {
@@ -166,7 +103,7 @@ export default class App extends Component<Props, State> {
     const version = process.env.REACT_APP_VERSION;
     const paths = createWhoOwnsWhatRoutePaths();
 
-
+/*
     let ToggleInfoButton = () => {
       const [showWidget, setShowWidget] = React.useState(false)
       const onClick = () => setShowWidget(true)
@@ -175,7 +112,7 @@ export default class App extends Component<Props, State> {
           { showWidget ? <div onClick={onClick} >
         Toggled
       </div> :  <div onClick={onClick} >
-        Toggled
+        Not Toggled
       </div> }
           { showWidget ? <Widget/> : null }
           
@@ -189,106 +126,7 @@ export default class App extends Component<Props, State> {
       </div>
     )
 
-
-    //header of the widget, says "What's New" and has close button
-    let widgetHeader = (
-      <div className="widget-header">
-        <span className="widget-title focusable" role="heading" tabIndex={0}>
-          What's New
-        </span>
-        <button
-          type="button"
-          className="widget-button-cancel material-icons md-18 focusable"
-          tabIndex={0}
-          onClick={(event) => {
-            event.preventDefault();
-            let widget = document.getElementById("widget");
-            return this.toggleWidget(widget);
-          }}
-        >
-          close
-        </button>
-      </div>
-    );
-
-    //All entries in the widget will have this format: index, title, image, and description
-    let widgetEntries = this.state.entries.map((entry) => (
-      <div
-        className="widget-entry"
-        id={"page-" + entry.title}
-        role="region"
-        aria-labelledby={"widget-" + entry.title + "-id"}
-      >
-        <p className="widget-entry-title-container">
-          <span className="widget-entry-index" aria-describedby={entry.index}>
-            {entry.index}
-          </span>
-          <span
-            className="widget-entry-title"
-            aria-describedby={"widget-entry title- " + entry.title}
-          >
-            {entry.title}
-          </span>
-        </p>
-        <img className="widget-entry-image" src={entry.img} alt={entry.description}></img>
-        <p className="widget-entry-description">{entry.description}</p>
-      </div>
-    ));
-
-    //Prev/next buttons
-    let navButtons = (
-      <div className="widget-nav-buttons-container">
-        <button
-          className="widget-button-nav prev focusable"
-          tabIndex={0}
-          onClick={(event) => {
-            event.preventDefault();
-            return this.prevEntry();
-          }}
-        >
-          <span className="material-icons md-14 widget-prev-next-icon">navigate_before</span>
-          <span className="widget-prev-text">Prev</span>
-        </button>
-        <button
-          className="widget-button-nav next focusable"
-          tabIndex={0}
-          onClick={(event) => {
-            event.preventDefault();
-            return this.nextEntry();
-          }}
-        >
-          <span className="widget-next-text">Next</span>
-          <span className="material-icons md-14 widget-prev-next-icon">navigate_next</span>
-        </button>
-      </div>
-    );
-
-    //The whole widget, with the info button toggle
-    let featureCalloutWidget = (
-      <div>
-        <div className="widget-triangle-info-button-container">
-          <button
-            className="widget-button-info focusable material-icons md-14"
-            tabIndex={0}
-            onClick={(event) => {
-              event.preventDefault();
-              let widget = document.getElementById("widget");
-              return this.toggleWidget(widget);
-            }}
-          >
-            info
-          </button>
-          <div className="widget-tooltip-triangle"></div>
-        </div>
-        <div className="widget-container" id="widget">
-          {widgetHeader}
-          <div className="widget-content-container" id="widget-entries">
-            {widgetEntries[this.state.entryIndex]}
-            {navButtons}
-          </div>
-        </div>
-      </div>
-    );
+*/
 
     return (
       <Router>
@@ -316,8 +154,7 @@ export default class App extends Component<Props, State> {
                   </span>
                 )}
                 <nav className="inline">
-                  {featureCalloutWidget}
-                  {ToggleInfoButton}
+                  <FeatureCalloutWidget/>
                   <LocaleNavLink exact to={paths.home}>
                     <Trans>Home</Trans>
                   </LocaleNavLink>
